webScrapping/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
anyio==3.6.2
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.2.0
auto-py-to-exe==2.36.0
beautifulsoup4==4.9.3
bidict==0.22.1
bottle==0.12.25
bottle-websocket==0.2.9
branca==0.6.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
cv2-tools==2.4.0
Django==4.1.7
EasyProcess==1.1
Eel==0.16.0
entrypoint2==1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.0
fastapi==0.95.0
filetype==1.2.0
Flask==2.2.3
Flask-SocketIO==5.3.4
flatbuffers==23.3.3
folium==0.14.0
frozenlist==1.3.3
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
googlesearch-python==1.1.0
greenlet==2.0.2
h11==0.14.0
humanfriendly==10.0
idna==2.10
ImageHash==4.3.1
imageio==2.27.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
lazy_loader==0.2
llvmlite==0.39.1
lxml==4.9.2
MarkupSafe==2.1.2
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
multidict==6.0.4
mysql-connector-python==8.1.0
networkx==3.1
numba==0.56.4
numpy==1.23.5
onnxruntime==1.14.1
openai==0.27.6
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
openpyxl==3.0.10
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pefile==2023.2.7
phonenumbers==8.13.16
Pillow==9.5.0
platformdirs==3.2.0
pooch==1.7.0
protobuf==4.21.12
py-mini-racer==0.6.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==1.10.7
PyGetWindow==0.0.9
pyinstaller==5.12.0
pyinstaller-hooks-contrib==2023.3
PyMatting==1.1.8
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyreadline3==3.4.1
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
pytesseract==0.3.10
python-constraint==1.4.0
python-dateutil==2.8.2
python-engineio==4.4.1
python-multipart==0.0.6
python-socketio==5.8.0
pytweening==1.0.7
pytz==2022.7.1
PyWavelets==1.4.1
pywin32-ctypes==0.2.0
rembg==2.0.32
requests==2.31.0
scapy==2.5.0
scikit-image==0.20.0
scipy==1.10.1
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
starlette==0.26.1
sympy==1.11.1
tabulate==0.9.0
tifffile==2023.3.21
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.9.2
truecallerpy==0.1.2
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
uvicorn==0.21.1
watchdog==3.0.0
Werkzeug==2.2.3
whichcraft==0.6.1
wsproto==1.2.0
wxPython==4.2.1
xlrd==2.0.1
yarl==1.8.2
zope.event==4.6
zope.interface==6.0
